feat: Add format for later on
The `format` attribute serves as a hint for how to display the question.
It is a specially-formatted string. It -- coupled with the type -- tells
the view how to render a given question.

For instance, the multiple choice question will have a format like this:

foo|bar|baz

This means that it should have three choices called 'foo', 'bar' and
'baz' for the user to pick.